Minor cleanup of role-mining feature

+++ gui/admin-gui/src/main/java/com/evolveum/midpoint/gui/impl/page/admin/role/mining/algorithm/utils/Handler.java
@@ -55,7 +55,7 @@ public class Handler implements Serializable {
     public void iterateActualStatus() {
         this.actualStatus++;
         this.endTime = System.currentTimeMillis();
-        this.percentage = (actualStatus / operationCountToProcess) * 100;
+        this.percentage = (int) ((actualStatus / (double) operationCountToProcess) * 100);
         print();
     }
